requirements.txt
argh==0.29.4
asgiref==3.7.2
autodiscover==0.0.4
baton==1.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
decouple==0.0.7
Django==4.2.6
django-baton==2.8.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docker-py==1.9.0
drf-yasg==1.21.7
hgicommon==1.2.0
hgijson==1.3.1
idna==3.4
inflection==0.5.1
packaging==23.2
pathtools==0.1.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.6
watchdog==0.8.3
websocket-client==1.6.4
